+++ Functions/nonContinuousRastrigin.h
@@ -11,9 +11,9 @@
 	#define MIN_SEARCH_RANGE -5.12f
 	#define MAX_SEARCH_RANGE 5.12f
 
-	float formulae(float values[]);
-		float calculateValue(float value);
-			int round(float number);
+	double formulae(double values[]);
+		double calculateValue(double value);
+			int round(double number);
 
 #endif
 #endif
